split huge jmeter.py, separate jmx into package
