+++ bzt/jmx/__init__.py
@@ -1,0 +1,2 @@
+from bzt.jmx.jmx import JMX
+from bzt.jmx.tools import JMeterScenarioBuilder, LoadSettingsProcessor
