+++ bzt/jmx/jmx.py
@@ -1071,132 +1071,3 @@ class JMX(object):
         action.append(JMX.int_prop("ActionProcessor.target", target_index))
         action.append(JMX._string_prop("ActionProcessor.duration", str(duration_ms)))
         return action
-
-
-class AbstractThreadGroup(object):
-    XPATH = None
-
-    def __init__(self, element, logger):
-        self.element = element
-        self.gtype = self.__class__.__name__
-        self.log = logger.getChild(self.gtype)
-
-    def create(self):  # todo: delegate content of group creation to itself?
-        return None
-
-    def get_concurrency(self):
-        self.log.warning("Getting of concurrency isn't suported for %s, choose 1" % self.gtype)
-        return 1
-
-    def get_action_on_error(self):
-        action = self.element.find(".//stringProp[@name='ThreadGroup.on_sample_error']")
-        if action is not None:
-            return action.text
-
-
-class ThreadGroup(AbstractThreadGroup):
-    XPATH = 'jmeterTestPlan>hashTree>hashTree>ThreadGroup'
-    CONCURRENCY_SEL = ".//stringProp[@name='ThreadGroup.num_threads']"
-
-    def get_concurrency(self):
-        concurrency_element = self.element.find(self.CONCURRENCY_SEL)
-        try:
-            concurrency = int(concurrency_element.text)
-        except ValueError:
-            msg = "Parsing concurrency '%s' in group '%s' failed, choose 1"
-            self.log.warning(msg, concurrency_element.text, self.gtype)
-            concurrency = 1
-
-        return concurrency
-
-
-class SteppingThreadGroup(AbstractThreadGroup):
-    XPATH = r'jmeterTestPlan>hashTree>hashTree>kg\.apc\.jmeter\.threads\.SteppingThreadGroup'
-    CONCURRENCY_SEL = ".//stringProp[@name='ThreadGroup.num_threads']"
-
-    def get_concurrency(self):
-        concurrency_element = self.element.find(self.CONCURRENCY_SEL)
-        try:
-            concurrency = int(concurrency_element.text)
-        except ValueError:
-            msg = "Parsing concurrency '%s' in group '%s' failed, choose 1"
-            self.log.warning(msg, concurrency_element.text, self.gtype)
-            concurrency = 1
-
-        return concurrency
-
-
-class UltimateThreadGroup(AbstractThreadGroup):
-    XPATH = r'jmeterTestPlan>hashTree>hashTree>kg\.apc\.jmeter\.threads\.UltimateThreadGroup'
-
-
-class ConcurrencyThreadGroup(AbstractThreadGroup):
-    XPATH = r'jmeterTestPlan>hashTree>hashTree>com\.blazemeter\.jmeter\.threads\.concurrency\.ConcurrencyThreadGroup'
-
-
-class ThreadGroupHandler(object):
-    CLASSES = [ThreadGroup, SteppingThreadGroup, UltimateThreadGroup, ConcurrencyThreadGroup]
-
-    def __init__(self, logger):
-        self.log = logger.getChild(self.__class__.__name__)
-
-    def groups(self, jmx):
-        """
-        Get wrappers for thread groups that are enabled
-        """
-        for _class in self.CLASSES:
-            for group in jmx.get(_class.XPATH):
-                if group.get("enabled") != 'false':
-                    yield _class(group, self.log)
-
-    def convert2tg(self, group, load, concurrency):
-        """
-        Convert TGs to simple ThreadGroup for load applying
-        """
-        testname = group.element.get('testname')
-        self.log.warning("Converting %s (%s) to normal ThreadGroup", group.gtype, testname)
-        on_error = group.get_action_on_error()
-        iterations = load.iterations
-        if not iterations:
-            iterations = -1
-
-        new_group_element = JMX.get_thread_group(
-            concurrency=concurrency,
-            iterations=iterations,
-            rampup=load.ramp_up,
-            testname=testname,
-            on_error=on_error)
-
-        # add scheduler if necessary
-        if load.hold or (load.ramp_up and not load.iterations):
-            sched_sel = "[name='ThreadGroup.scheduler']"
-            sched_xpath = GenericTranslator().css_to_xpath(sched_sel)
-            dur_sel = "[name='ThreadGroup.duration']"
-            dur_xpath = GenericTranslator().css_to_xpath(dur_sel)
-
-            new_group_element.xpath(sched_xpath)[0].text = 'true'
-            new_group_element.xpath(dur_xpath)[0].text = str(int(load.duration))
-            loops_element = new_group_element.find(".//elementProp[@name='ThreadGroup.main_controller']")
-            loops_loop_count = loops_element.find("*[@name='LoopController.loops']")
-            loops_loop_count.getparent().replace(loops_loop_count, JMX.int_prop("LoopController.loops", -1))
-
-        group.element.getparent().replace(group.element, new_group_element)
-
-    def convert2ctg(self, group, load, concurrency):
-        """
-        Convert TGs to ConcurrencyThreadGroup for load applying
-        """
-        testname = group.element.get('testname')  # todo: add to group interface
-        self.log.warning("Converting %s (%s) to ConcurrencyThreadGroup", group.gtype, testname)
-        on_error = group.get_action_on_error()
-        iterations = load.iterations
-
-        # todo: add steps and iterations
-        new_group_element = JMX.get_concurrency_thread_group(
-            concurrency=concurrency,
-            rampup=load.ramp_up,
-            hold_for=load.hold,
-            on_error=on_error,
-            testname=testname)
-
-        group.element.getparent().replace(group.element, new_group_element)
